+++ plugin-gradle/src/main/java/dev/equo/ide/gradle/P2ModelDslWithCatalog.java
@@ -19,7 +19,6 @@ import dev.equo.ide.IdeHook;
 import dev.equo.ide.IdeHookBuildship;
 import java.io.File;
 import java.util.List;
-import org.gradle.api.Action;
 import org.gradle.api.GradleException;
 import org.gradle.api.Project;
 
@@ -34,6 +33,21 @@ public class P2ModelDslWithCatalog extends P2ModelDsl {
 		public Platform(String urlOverride) {
 			super(Catalog.PLATFORM, urlOverride);
 		}
+
+		public Platform showLineNumbers(boolean showLineNumbers) {
+			Catalog.PLATFORM.showLineNumbers(workspaceInit(), showLineNumbers);
+			return this;
+		}
+
+		public Platform showWhitespace(boolean showWhitespace) {
+			Catalog.PLATFORM.showWhitespace(workspaceInit(), showWhitespace);
+			return this;
+		}
+
+		public Platform showLineEndings(boolean showLineEndings) {
+			Catalog.PLATFORM.showLineEndings(workspaceInit(), showLineEndings);
+			return this;
+		}
 	}
 
 	public Platform platform(String urlOverride) {
@@ -47,6 +61,11 @@ public class P2ModelDslWithCatalog extends P2ModelDsl {
 	public static class Jdt extends GradleCatalogDsl {
 		public Jdt(String urlOverride) {
 			super(Catalog.JDT, urlOverride);
+		}
+
+		public Jdt classpathVariable(String name, String value) {
+			Catalog.JDT.classpathVariable(workspaceInit(), name, value);
+			return this;
 		}
 	}
 
@@ -116,15 +135,7 @@ public class P2ModelDslWithCatalog extends P2ModelDsl {
 	}
 
 	public Pde pde() {
-		return pde((String) null);
-	}
-
-	public Pde pde(String urlOverride, Action<Pde> dsl) {
-		return add(new Pde(urlOverride, project), dsl);
-	}
-
-	public Pde pde(Action<Pde> dsl) {
-		return pde(null, dsl);
+		return pde(null);
 	}
 
 	public static class M2E extends GradleCatalogDsl {
@@ -219,12 +230,6 @@ public class P2ModelDslWithCatalog extends P2ModelDsl {
 	}
 
 	private <T extends GradleCatalogDsl> T add(T dsl) {
-		add(dsl, unused -> {});
-		return dsl;
-	}
-
-	private <T extends GradleCatalogDsl> T add(T dsl, Action<? super T> action) {
-		action.execute(dsl);
 		catalog.add(dsl);
 		return dsl;
 	}