+++ plugin-maven/src/main/java/dev/equo/ide/maven/AbstractP2MojoWithCatalog.java
@@ -21,6 +21,7 @@ import dev.equo.ide.WorkspaceInit;
 import dev.equo.solstice.p2.P2Model;
 import java.io.File;
 import java.util.List;
+import java.util.Map;
 import java.util.Objects;
 import java.util.stream.Stream;
 import org.apache.maven.plugins.annotations.Parameter;
@@ -33,6 +34,29 @@ public abstract class AbstractP2MojoWithCatalog extends AbstractP2Mojo {
 		public Platform() {
 			super(Catalog.PLATFORM);
 		}
+
+		@Parameter(required = false)
+		private Boolean showLineNumbers;
+
+		@Parameter(required = false)
+		private Boolean showWhitespace;
+
+		@Parameter(required = false)
+		private Boolean showLineEndings;
+
+		@Override
+		protected void processVersionOverrides() {
+			super.processVersionOverrides();
+			if (showLineNumbers != null) {
+				Catalog.PLATFORM.showLineNumbers(workspaceInit(), showLineNumbers);
+			}
+			if (showWhitespace != null) {
+				Catalog.PLATFORM.showWhitespace(workspaceInit(), showWhitespace);
+			}
+			if (showLineEndings != null) {
+				Catalog.PLATFORM.showLineEndings(workspaceInit(), showLineEndings);
+			}
+		}
 	}
 
 	@Parameter private Jdt jdt;
@@ -40,6 +64,18 @@ public abstract class AbstractP2MojoWithCatalog extends AbstractP2Mojo {
 	public static class Jdt extends MavenCatalogDsl {
 		public Jdt() {
 			super(Catalog.JDT);
+		}
+
+		@Parameter(required = false)
+		private Map<String, String> classpathVariable;
+
+		@Override
+		protected void processVersionOverrides() {
+			super.processVersionOverrides();
+			if (classpathVariable != null) {
+				classpathVariable.forEach(
+						(name, value) -> Catalog.JDT.classpathVariable(workspaceInit(), name, value));
+			}
 		}
 	}