+++ plugin-maven/CHANGELOG.md
@@ -18,12 +18,13 @@ We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format.
     <configuration>
       <workspaceProps>
         <workspaceProp>
-            <path>.metadata/.plugins/org.eclipse.core.runtime/.settings/org.eclipse.pde.api.tools.prefs</path>
+            <path>instance/.metadata/.plugins/org.eclipse.core.runtime/.settings/org.eclipse.pde.api.tools.prefs</path>
             <key>missing_default_api_profile</key>
             <value>Ignore</value>
         </workspaceProp>
         ...
     ```
+- Control preferences for whitespace and line numbers under `platform`, add classpath variables under `jdt`. ([#128](https://github.com/equodev/equo-ide/pull/128))
 ### Fixed
 - Eclipse PDE now runs well under Atomos. ([#126](https://github.com/equodev/equo-ide/pull/126))